soal_2/client_1/requirements.txt
djangorestframework
djangorestframework-simplejwt==5.3.1
python-dotenv
